requirements.txt
**`requirements.txt`:**
```
gradio
sentence-transformers
scikit-learn
PyPDF2
pdfplumber
pdf2image
pytesseract
Pillow
numpy
pandas
scipy
python-dateutil
torch
```

**`packages.txt`:**
```
tesseract-ocr
poppler-utils